terminate simulation correctly!
Former-commit-id: bc85b5c17969a0b15a1cd92d18e585cfa032c924
Former-commit-id: 4bccd55157f20f9952763490233787cdcd420f66

+++ application/traci/TraCI_Start.h
@@ -42,9 +42,9 @@ public:
     std::string routeId;
     double pos;
     double speed;
-    uint8_t lane;
+    uint32_t lane;
 
-    departedNodes(std::string str1, std::string str2, std::string str3, double d1, double d2, uint8_t i2)
+    departedNodes(std::string str1, std::string str2, std::string str3, double d1, double d2, uint32_t i2)
     {
         this->vehicleId = str1;
         this->vehicleTypeId = str2;
